Add feature and step id, hidden field to Cucumber json report
