+++ src/Pickles/Pickles/Extensions/StringExtensions.cs
@@ -19,13 +19,9 @@
 //  --------------------------------------------------------------------------------------------------------------------
 
 using System;
-using System.Linq;
-using System.Text;
 
 namespace PicklesDoc.Pickles.Extensions
 {
-    using System.Text.RegularExpressions;
-
     public static class StringExtensions
     {
         /// <summary>
@@ -41,28 +37,6 @@ namespace PicklesDoc.Pickles.Extensions
                     .Replace("\r", string.Empty)
                     .Replace("\n", Environment.NewLine)
                     .Replace("\t", "    ");
-        }
-
-        /// <summary>
-        /// Takes a string and returns a url-friendly version of the string (slug) with all special characters
-        /// and extra spaces stripped out, with words seperated by dashes.
-        /// </summary>
-        /// <param name="text">The string that will be used to create the slug.</param>
-        /// <returns>A slug generated from the given string.</returns>
-        public static string ToSlug(this string text)
-        {
-            // remove any accent characters
-            Encoding.RegisterProvider(CodePagesEncodingProvider.Instance);
-            var bytes = Encoding.GetEncoding("Cyrillic").GetBytes(text);
-            var str = Encoding.ASCII.GetString(bytes);
-
-            // modify string to slug format
-            str = str.ToLower();
-            str = Regex.Replace(str, @"[^a-z0-9\s-]", "");
-            str = Regex.Replace(str, @"\s+", " ").Trim();
-            str = Regex.Replace(str, @"\s", "-");
-
-            return str;
         }
     }
 }
